+++ include/string.h
@@ -8,6 +8,7 @@ int strcmp(const char* s1, const char* s2);
 int strncmp(const char* s1, const char* s2, size_t n);
 char* strcpy(char* dest, const char* src);
 char* strncpy(char* dest, const char* src, size_t n);
+char* strcat(char* dest, const char* src);
 void* memset(void* ptr, int value, size_t num);
 void* memcpy(void* dest, const void* src, size_t n);