+++ kernel/lib/string.c
@@ -42,6 +42,15 @@ char* strncpy(char* dest, const char* src, size_t n) {
     return ret;
 }
 
+char* strcat(char* dest, const char* src) {
+    char* ret = dest;
+    while (*dest)
+        dest++;
+    while ((*dest++ = *src++))
+        ;
+    return ret;
+}
+
 void* memset(void* ptr, int value, size_t num) {
     unsigned char* p = ptr;
     while (num--)